Ensuring that all hash values get set.
Previously if you set one value, it would not try to set the others as
id? would return true, since the hash was not blank once it had a
single key. This overrides id? to ensure that it returns false if any
of the values are blank.

Really need to introduce a new type here...