Added json type.
Encodes and decodes using active supports json. Does not try to decode
non-string values. This allows you to stay in ruby land until to_store.
from_store is called whenever a value is assigned in toystore so
without this, you would always have to assign a json string. Not fun.